# huff0: asm implementation of Decompress1X