@klauspost 2022-05-23
Nice! 

BMI doesn't show any speedup on my system, but also no worse.